// src/BindingChaos.Web/src/api/models/SocietyListItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Response model for a society in a list.
 * @export
 * @interface SocietyListItemResponse
 */
export interface SocietyListItemResponse {
    /**
     * The unique identifier of the society.
     * @type {string}
     * @memberof SocietyListItemResponse
     */
    id?: string;
    /**
     * The name of the society.
     * @type {string}
     * @memberof SocietyListItemResponse
     */
    name?: string;
    /**
     * The description of the society.
     * @type {string}
     * @memberof SocietyListItemResponse
     */
    description?: string;
    /**
     * When the society was created.
     * @type {Date}
     * @memberof SocietyListItemResponse
     */
    createdAt?: Date;
    /**
     * The tags associated with the society.
     * @type {Array<string>}
     * @memberof SocietyListItemResponse
     */
    tags?: Array<string>;
    /**
     * Whether this society has geographic bounds.
     * @type {boolean}
     * @memberof SocietyListItemResponse
     */
    hasGeographicBounds?: boolean;
    /**
     * The number of active members.
     * @type {number}
     * @memberof SocietyListItemResponse
     */
    activeMemberCount?: number;
}

/**
 * Check if a given object implements the SocietyListItemResponse interface.
 */
export function instanceOfSocietyListItemResponse(value: object): value is SocietyListItemResponse {
    return true;
}

export function SocietyListItemResponseFromJSON(json: any): SocietyListItemResponse {
    return SocietyListItemResponseFromJSONTyped(json, false);
}

export function SocietyListItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SocietyListItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'tags': json['tags'] == null ? undefined : json['tags'],
        'hasGeographicBounds': json['hasGeographicBounds'] == null ? undefined : json['hasGeographicBounds'],
        'activeMemberCount': json['activeMemberCount'] == null ? undefined : json['activeMemberCount'],
    };
}

export function SocietyListItemResponseToJSON(json: any): SocietyListItemResponse {
    return SocietyListItemResponseToJSONTyped(json, false);
}

export function SocietyListItemResponseToJSONTyped(value?: SocietyListItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'tags': value['tags'],
        'hasGeographicBounds': value['hasGeographicBounds'],
        'activeMemberCount': value['activeMemberCount'],
    };
}